refactor(proxy): split the pass-through attempt out of multimodal.go
The saturation-retry branch pushed multimodal.go one line over the 800-line
size ceiling, and the gate takes no new allowlist entries: the per-candidate
attempt (attemptPassthroughCandidate + passthroughAnswered) moves whole into
multimodal_attempt.go, mirroring how the chat attempt lives beside its serve
pipeline rather than inside it. No behavioural change.
